move Header into it's own component
